use methods remove from crowi dependence
